hl_scraper/requirements.txt
requests
pandas
psycopg2-binary
python-dotenv
Flask
flask-cors
